Added game logic
Added game logic; if a player places a cross, a nought is put in a random position. This will in future be an AI that decides the best move. Added winning, losing and drawing logic. Added reset logic, so after a game has been won, lost or drawn the game resets.