Added delay after cross placed
When the user places the cross, the cross renders straight away, but waits one second, then the nought is placed. This makes the game a lot more realistic and smoother. Within this one second another cross cannot be placed. This works using threads. There is an "update" thread and a "render" thread. The game loop is the same as before, but with multitasking class methods. When a cross is placed, the update thread waits 1 second, while the rendering thread places the cross. The next frame is not loaded untill the 1 second wait in the "update" thread has finished, so the next frame is when the nought is rendered.