Added minimax algorithm for AI
Added the minimax algorithm to search the best moves for the computer. This is currently not finished as the computer doesn't always play the best moves. This is because the depth is currently set to 2. It would need to be set to 9 (Maximum number of moves to look in the future) to optimise the AI, but currently there are some issues where it seems to play worse with a higher depth. I think this is something to do with how I'm storing the index for the best position when it is found - but this needs looking into.